requirements.txt
flask==2.0.1
python-docx==0.8.11
nltk==3.6.3
werkzeug==2.0.1
gunicorn==20.1.0
